fix: Streamlit Cloud 배포 대응
- requirements.txt 추가 (없으면 scikit-learn 미설치로 배포 실패)
- 원-핫 인코딩을 int8 로 변경해 메모리 85MB -> 15MB (예측값 동일 확인)

+++ app.py
@@ -58,7 +58,9 @@ def build_model() -> dict:
     """데이터를 읽고 나무를 학습한다. 앱 실행 중 한 번만 수행된다."""
     df = pd.read_csv(SPLIT_PATH)
     dummies = pd.get_dummies(df[CATEGORICAL], columns=CATEGORICAL, drop_first=False)
-    X = pd.concat([dummies.astype(int), df[NUMERIC]], axis=1)
+    # int8 로 두면 원-핫 행렬이 85MB -> 15MB 로 줄어든다.
+    # sklearn 이 학습 시 내부적으로 float32 로 변환하므로 예측값은 int 와 완전히 동일하다.
+    X = pd.concat([dummies.astype("int8"), df[NUMERIC]], axis=1)
 
     is_train = df[SCHEME] == "train"
     y_tr = df.loc[is_train, TARGET].to_numpy(float)
